move data folder into src and modify references to it
